# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: UN17 Village Loppemarked API
  version: 0.1.0
  description: >
    Contract-first API skeleton for UN17 Village Loppemarked registration and admin operations.
servers:
  - url: https://loppemarked.un17hub.com
...
paths:
...
  /public/register:
    post:
      tags: [Public]
      summary: Register an apartment for a flea-market table.
      responses:
        '200':
          description: Registration accepted.
        '409':
          description: Apartment already has registration (switch confirmation needed).
...
  /admin/auth/login:
    post:
      tags: [Admin]
      summary: Admin login with email and password.
      responses:
        '200':
          description: Authenticated.
        '401':
          description: Invalid credentials.
...
  /admin/waitlist/{id}:
    delete:
      tags: [Admin]
      summary: Remove a waitlist entry without assigning a table.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Removed.
...
  /admin/admins/{id}:
    delete:
      tags: [Admin]
      summary: Delete admin account (except self).
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted.
...